// frontend/src/types/responses/categories_response.d.ts
export interface CategoriesResponse {
    data: Daum[]
    meta: Meta
}

export interface Daum {
    id: number
    attributes: Attributes
}

export interface Attributes {
    name: string
    createdAt: string
    updatedAt: string
    publishedAt: string
    image: Image
    products: Products
    icon: Icon
}

export interface Image {
    data: Data
}

export interface Data {
    id: number
    attributes: Attributes2
}

export interface Attributes2 {
    name: string
    alternativeText: any
    caption: any
    width: number
    height: number
    formats: Formats
    hash: string
    ext: string
    mime: string
    size: number
    url: string
    previewUrl: any
    provider: string
    provider_metadata: any
    createdAt: string
    updatedAt: string
}

export interface Formats {
    large: Large
    small: Small
    medium: Medium
    thumbnail: Thumbnail
}

export interface Large {
    ext: string
    url: string
    hash: string
    mime: string
    name: string
    path: any
    size: number
    width: number
    height: number
}

export interface Small {
    ext: string
    url: string
    hash: string
    mime: string
    name: string
    path: any
    size: number
    width: number
    height: number
}

export interface Medium {
    ext: string
    url: string
    hash: string
    mime: string
    name: string
    path: any
    size: number
    width: number
    height: number
}

export interface Thumbnail {
    ext: string
    url: string
    hash: string
    mime: string
    name: string
    path: any
    size: number
    width: number
    height: number
}

export interface Products {
    data: Daum2[]
}

export interface Daum2 {
    id: number
    attributes: Attributes3
}

export interface Attributes3 {
    title: string
    is_available: boolean
    original_price: number
    discount: number
    short_description: string
    long_description: string
    createdAt: string
    updatedAt: string
    publishedAt: string
}

export interface Icon {
    data: Data2
}

export interface Data2 {
    id: number
    attributes: Attributes4
}

export interface Attributes4 {
    name: string
    alternativeText: any
    caption: any
    width: number
    height: number
    formats: any
    hash: string
    ext: string
    mime: string
    size: number
    url: string
    previewUrl: any
    provider: string
    provider_metadata: any
    createdAt: string
    updatedAt: string
}

export interface Meta {
    pagination: Pagination
}

export interface Pagination {
    page: number
    pageSize: number
    pageCount: number
    total: number
}